refactor(tau-events): split lib tests into module (#1375)
